+++ packages/i18n/scripts/build-npm.ts
@@ -2,23 +2,29 @@ import fs from 'node:fs';
 import path from 'node:path';
 import package_ from '../package.json';
 import ar from '../src/ar';
+import az from '../src/az';
 import ca from '../src/ca';
 import cs from '../src/cs';
 import de from '../src/de';
+import el from '../src/el';
 import es from '../src/es';
 import fa from '../src/fa';
+import fi from '../src/fi';
 import fr from '../src/fr';
 import hu from '../src/hu';
 import id from '../src/id';
 import it from '../src/it';
 import ja from '../src/ja';
+import ko from '../src/ko';
 import kr from '../src/kr';
+import mn from '../src/mn';
 import nb from '../src/nb';
 import nl from '../src/nl';
 import pl from '../src/pl';
 import pt from '../src/pt';
 import ro from '../src/ro';
 import ru from '../src/ru';
+import sk from '../src/sk';
 import sl from '../src/sl';
 import sv from '../src/sv';
 import tr from '../src/tr';
@@ -27,31 +33,34 @@ import vi from '../src/vi';
 import zhCN from '../src/zh-CN';
 import zhTW from '../src/zh-TW';
 
-// Start timer
-console.time('build');
-
 // Create languages array
 // Note: The language file `en` does not need to be added as the default
 // messages of Valibot are already in English
 const languages = [
   ar,
+  az,
   ca,
   cs,
   de,
+  el,
   es,
   fa,
+  fi,
   fr,
   hu,
   id,
   it,
   ja,
+  ko,
   kr,
+  mn,
   nb,
   nl,
   pl,
   pt,
   ro,
   ru,
+  sk,
   sl,
   sv,
   tr,
@@ -290,6 +299,3 @@ fs.writeFileSync(
 
 // Write root .gitignore file
 fs.writeFileSync('.gitignore', files.join('\n'));
-
-// End timer
-console.timeEnd('build');